Checksum offloading support.
vRouter advertises checksum offloadin capabilities to VM. If VM wants to use
the offloading, proper flag in packets is set. vRouter checks for the flag and
either computes checksum in software or does the offloading itself to the NIC
installed in compute node (if NIC supports hardware checksum computing).

The code contains functions ported from DPDK 1.8.0.

Change-Id: I2cfa47044c8f26051d62f0650536842fe33567e7